Route differently in prod and other environments
For now, we want the landing page in prod to be the about page, while in
the other environments we want the landing page to be the books
